+++ NEWS.md
@@ -1,4 +1,5 @@
 # dendroTools 1.0.0.
 
+* The use_median argument is removed and replaced with aggregate_function argument. This argument is a character string specifying how the daily data should be aggregated. The default is ‘mean’, the two other options are 'median' and 'sum'.
 * Additional arguments are added, holdout, holdout_share and holdout_manual
 * Package version is changed to 1.0.0